refactor(ErrorMessage): remove specialized error message components
