--- src/components/missingContentErrorMessage/missingContentErrorMessage.module.css
@@ -1,8 +1,0 @@
-.description {
-    font-family: monospace;
-    font-size: 1rem;
-    margin: 0.5rem 0;
-    background: var(--primary-white);
-    border-radius: 8px;
-    padding: 0.75rem;
-}
